2022 prep: add 2018 day 3
Part 1: 5:23 (with one wrong answer)
Part 2: 7:15
Would make it in top 50 that year

Very happy with the fact that I used solution to part 1 in part 2
instead of trying to be super smart and solving it in some 'optimal
way'. It's probably a bit ugly, but it works and it's fast so who cares.