Add two more solutions to day 15
One is just a Z3 solver to which you insert the inequalities and I wish I knew
about this solver when I was doing the exercise itself.

The second one goes around the border of each sensor as the solution must be on
a border of at least one of them. Otherwise, we could move it and it would
still be a solution.

Z3 solves it almost instantly, the other two solutions take some time.