Add day 19
What a carnage lol
Part 1: 3:51:32 (2303rd)
Part 2: 7:33:56 (2983rd)

Implemented in C++ (first with no optimisations a solution that ran in like 30
minutes). Accepts input as the space-separated numbers indicating the costs as
given in the original input. Runs in like 10 seconds maybe. Still not sure if
the greedy approach of always getting the geode robot and if possible obsidian
robot is mathematically proof or just input+minuts proof as with these inputs
and for these small amounts of minutes it will just always work.