Add 2023 day 20 part 2
Along with the 'real', 'behind the scenes' manual solution using graph plotting
in Python. This was actually super helpful and the Haskell implementation is
just to complete this work. I would expect my solution to work on all curated
inputs for this puzzle hopefully, although it is in no way generic.