Polish 2023 days 21 and 22
21 is still disgusting and after having a look at the subreddit it should have
been way simpler. Like not even noticing that it's quadratic, but maybe
visualising and seeing that it forms a perfect 'diamond' that repeats, in
consequence giving us a nicely growing series. This depends on many assumptions
about the input (no obstacles in the middle, horizontally and vertically; all
squares in the diamond are reachable at least as quickly as the corners), but I
did notice the first one so I guess the lesson is to look deeper if the
solution seems too complicated instead of trying to write the most complicated
algorithm possible.